virtio_spec::pci

Struct NotificationData

pub struct NotificationData(/* private fields */);

Notification Data.

Implementations

impl NotificationData

pub const fn new() -> Self

Creates a new default initialized bitfield.

pub const fn from_bits(bits: le32) -> Self

Convert from bits.

pub const fn into_bits(self) -> le32

Convert into bits.

pub const fn vqn(&self) -> u16

VQ number to be notified.

Bits: 0..16

pub const fn with_vqn_checked(self, value: u16) -> Result<Self, ()>

VQ number to be notified.

Bits: 0..16

pub const fn with_vqn(self, value: u16) -> Self

VQ number to be notified.

Bits: 0..16

pub fn set_vqn(&mut self, value: u16)

VQ number to be notified.

Bits: 0..16

pub fn set_vqn_checked(&mut self, value: u16) -> Result<(), ()>

VQ number to be notified.

Bits: 0..16

pub const fn next_off(&self) -> u16

Offset within the ring where the next available ring entry will be written. When VIRTIO_F_RING_PACKED has not been negotiated this refers to the 15 least significant bits of the available index. When VIRTIO_F_RING_PACKED has been negotiated this refers to the offset (in units of descriptor entries) within the descriptor ring where the next available descriptor will be written.

Bits: 16..31

pub const fn with_next_off_checked(self, value: u16) -> Result<Self, ()>

Offset within the ring where the next available ring entry will be written. When VIRTIO_F_RING_PACKED has not been negotiated this refers to the 15 least significant bits of the available index. When VIRTIO_F_RING_PACKED has been negotiated this refers to the offset (in units of descriptor entries) within the descriptor ring where the next available descriptor will be written.

Bits: 16..31

pub const fn with_next_off(self, value: u16) -> Self

Offset within the ring where the next available ring entry will be written. When VIRTIO_F_RING_PACKED has not been negotiated this refers to the 15 least significant bits of the available index. When VIRTIO_F_RING_PACKED has been negotiated this refers to the offset (in units of descriptor entries) within the descriptor ring where the next available descriptor will be written.

Bits: 16..31

pub fn set_next_off(&mut self, value: u16)

Offset within the ring where the next available ring entry will be written. When VIRTIO_F_RING_PACKED has not been negotiated this refers to the 15 least significant bits of the available index. When VIRTIO_F_RING_PACKED has been negotiated this refers to the offset (in units of descriptor entries) within the descriptor ring where the next available descriptor will be written.

Bits: 16..31

pub fn set_next_off_checked(&mut self, value: u16) -> Result<(), ()>

Offset within the ring where the next available ring entry will be written. When VIRTIO_F_RING_PACKED has not been negotiated this refers to the 15 least significant bits of the available index. When VIRTIO_F_RING_PACKED has been negotiated this refers to the offset (in units of descriptor entries) within the descriptor ring where the next available descriptor will be written.

Bits: 16..31

pub const fn next_wrap(&self) -> u8

Wrap Counter. With VIRTIO_F_RING_PACKED this is the wrap counter referring to the next available descriptor. Without VIRTIO_F_RING_PACKED this is the most significant bit (bit 15) of the available index.

Bits: 31..32

pub const fn with_next_wrap_checked(self, value: u8) -> Result<Self, ()>

Wrap Counter. With VIRTIO_F_RING_PACKED this is the wrap counter referring to the next available descriptor. Without VIRTIO_F_RING_PACKED this is the most significant bit (bit 15) of the available index.

Bits: 31..32

pub const fn with_next_wrap(self, value: u8) -> Self

Wrap Counter. With VIRTIO_F_RING_PACKED this is the wrap counter referring to the next available descriptor. Without VIRTIO_F_RING_PACKED this is the most significant bit (bit 15) of the available index.

Bits: 31..32

pub fn set_next_wrap(&mut self, value: u8)

Wrap Counter. With VIRTIO_F_RING_PACKED this is the wrap counter referring to the next available descriptor. Without VIRTIO_F_RING_PACKED this is the most significant bit (bit 15) of the available index.

Bits: 31..32

pub fn set_next_wrap_checked(&mut self, value: u8) -> Result<(), ()>

Wrap Counter. With VIRTIO_F_RING_PACKED this is the wrap counter referring to the next available descriptor. Without VIRTIO_F_RING_PACKED this is the most significant bit (bit 15) of the available index.

Bits: 31..32

impl NotificationData

pub const fn next_idx(&self) -> u16

Available index

This collides with Self::next_off and Self::next_wrap.

Bits: 16..32

pub const fn with_next_idx(self, value: u16) -> Self

Available index

This collides with Self::with_next_off and Self::with_next_wrap.

Bits: 16..32

pub fn set_next_idx(&mut self, value: u16)

Available index

This collides with Self::set_next_off and Self::set_next_wrap.

Bits: 16..32