virtio_spec::volatile

Struct WideVolatilePtr

pub struct WideVolatilePtr<'a, T, A>
where
    T: ?Sized,
{ /* private fields */ }

A wide volatile pointer for 64-bit fields.

In virtio, 64-bit fields are to be treated as two 32-bit fields, with low 32 bit part followed by the high 32 bit part. This type mimics VolatilePtr, and allows easy access to 64-bit fields.

Implementations

impl<'a, T, A> WideVolatilePtr<'a, T, A>

pub fn from_low_high( low: VolatilePtr<'a, T, A>, high: VolatilePtr<'a, T, A>, ) -> Self

Creates a new wide volatile pointer from pointers to the low and to the high part.

impl<'a, A> WideVolatilePtr<'a, le32, A>

pub fn read(self) -> le64
where A: Readable,

Performs a volatile read of the contained value.

See VolatilePtr::read.

pub fn write(self, value: le64)
where A: Writable,

Performs a volatile write, setting the contained value to the given value.

See VolatilePtr::write.

pub fn update(self, f: impl FnOnce(le64) -> le64)
where A: Readable + Writable,

Updates the contained value using the given closure and volatile instructions.

See VolatilePtr::update.

impl<'a, A> WideVolatilePtr<'a, be32, A>

pub fn read(self) -> be64
where A: Readable,

Performs a volatile read of the contained value.

See VolatilePtr::read.

pub fn write(self, value: be64)
where A: Writable,

Performs a volatile write, setting the contained value to the given value.

See VolatilePtr::write.

pub fn update(self, f: impl FnOnce(be64) -> be64)
where A: Readable + Writable,

Updates the contained value using the given closure and volatile instructions.

See VolatilePtr::update.