virtio_spec::fs

Trait ConfigVolatileFieldAccess

pub trait ConfigVolatileFieldAccess<'a, A> {
    // Required methods
    fn tag(self) -> VolatilePtr<'a, [u8; 36], A::Restricted>
       where A: RestrictAccess<ReadOnly>;
    fn num_request_queues(self) -> VolatilePtr<'a, le32, A::Restricted>
       where A: RestrictAccess<ReadOnly>;
    fn notify_buf_size(self) -> VolatilePtr<'a, le32, A::Restricted>
       where A: RestrictAccess<ReadOnly>;
}

Device configuration

Use ConfigVolatileFieldAccess to work with this struct.

Required Methods

fn tag(self) -> VolatilePtr<'a, [u8; 36], A::Restricted>
where A: RestrictAccess<ReadOnly>,

This is the name associated with this file system. The tag is encoded in UTF-8 and padded with NUL bytes if shorter than the available space. This field is not NUL-terminated if the encoded bytes take up the entire field.

fn num_request_queues(self) -> VolatilePtr<'a, le32, A::Restricted>
where A: RestrictAccess<ReadOnly>,

This is the total number of request virtqueues exposed by the device. Each virtqueue offers identical functionality and there are no ordering guarantees between requests made available on different queues. Use of multiple queues is intended to increase performance.

fn notify_buf_size(self) -> VolatilePtr<'a, le32, A::Restricted>
where A: RestrictAccess<ReadOnly>,

This is the minimum number of bytes required for each buffer in the notification queue.