Struct virtio_spec::net::F

pub struct F(/* private fields */);

Network Device Feature Bits

Implementations

impl F

pub const CSUM: Self = _

Device handles packets with partial checksum. This “checksum offload” is a common feature on modern network cards.

pub const GUEST_CSUM: Self = _

Driver handles packets with partial checksum.

pub const CTRL_GUEST_OFFLOADS: Self = _

Control channel offloads reconfiguration support.

pub const MTU: Self = _

Device maximum MTU reporting is supported. If offered by the device, device advises driver about the value of its maximum MTU. If negotiated, the driver uses mtu as the maximum MTU value.

pub const MAC: Self = _

Device has given MAC address.

pub const GUEST_TSO4: Self = _

Driver can receive TSOv4.

pub const GUEST_TSO6: Self = _

Driver can receive TSOv6.

pub const GUEST_ECN: Self = _

Driver can receive TSO with ECN.

pub const GUEST_UFO: Self = _

Driver can receive UFO.

pub const HOST_TSO4: Self = _

Device can receive TSOv4.

pub const HOST_TSO6: Self = _

Device can receive TSOv6.

pub const HOST_ECN: Self = _

Device can receive TSO with ECN.

pub const HOST_UFO: Self = _

Device can receive UFO.

pub const MRG_RXBUF: Self = _

Driver can merge receive buffers.

pub const STATUS: Self = _

Configuration status field is available.

pub const CTRL_VQ: Self = _

Control channel is available.

pub const CTRL_RX: Self = _

Control channel RX mode support.

pub const CTRL_VLAN: Self = _

Control channel VLAN filtering.

pub const GUEST_ANNOUNCE: Self = _

Driver can send gratuitous packets.

pub const MQ: Self = _

Device supports multiqueue with automatic receive steering.

pub const CTRL_MAC_ADDR: Self = _

Set MAC address through control channel.

pub const HOST_USO: Self = _

Device can receive USO packets. Unlike UFO (fragmenting the packet) the USO splits large UDP packet to several segments when each of these smaller packets has UDP header.

pub const HASH_REPORT: Self = _

Device can report per-packet hash value and a type of calculated hash.

pub const GUEST_HDRLEN: Self = _

Driver can provide the exact hdr_len value. Device benefits from knowing the exact header length.

pub const RSS: Self = _

Device supports RSS (receive-side scaling) with Toeplitz hash calculation and configurable hash parameters for receive steering.

pub const RSC_EXT: Self = _

Device can process duplicated ACKs and report number of coalesced segments and duplicated ACKs.

pub const STANDBY: Self = _

Device may act as a standby for a primary device with the same MAC address.

pub const SPEED_DUPLEX: Self = _

Device reports speed and duplex.

pub const INDIRECT_DESC: Self = _

Device-independent Bit. See virtio::F::INDIRECT_DESC.

pub const EVENT_IDX: Self = _

Device-independent Bit. See virtio::F::EVENT_IDX.

pub const VERSION_1: Self = _

Device-independent Bit. See virtio::F::VERSION_1.

pub const ACCESS_PLATFORM: Self = _

Device-independent Bit. See virtio::F::ACCESS_PLATFORM.

pub const RING_PACKED: Self = _

Device-independent Bit. See virtio::F::RING_PACKED.

pub const IN_ORDER: Self = _

Device-independent Bit. See virtio::F::IN_ORDER.

pub const ORDER_PLATFORM: Self = _

Device-independent Bit. See virtio::F::ORDER_PLATFORM.

pub const SR_IOV: Self = _

Device-independent Bit. See virtio::F::SR_IOV.

pub const NOTIFICATION_DATA: Self = _

Device-independent Bit. See virtio::F::NOTIFICATION_DATA.

pub const NOTIF_CONFIG_DATA: Self = _

Device-independent Bit. See virtio::F::NOTIF_CONFIG_DATA.

pub const RING_RESET: Self = _

Device-independent Bit. See virtio::F::RING_RESET.

impl F

pub const fn empty() -> Self

Get a flags value with all bits unset.

pub const fn all() -> Self

Get a flags value with all known bits set.

pub const fn bits(&self) -> le128

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: le128) -> Option<Self>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: le128) -> Self

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: le128) -> Self

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<Self>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesn’t correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: Self) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: Self) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: Self)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: Self)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove won’t truncate other, but the ! operator will.

pub fn toggle(&mut self, other: Self)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: Self, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: Self) -> Self

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: Self) -> Self

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: Self) -> Self

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: Self) -> Self

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> Self

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl F

pub fn iter(&self) -> Iter<Self>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub fn iter_names(&self) -> IterNames<Self>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.