virtio_spec::pci

Trait CommonCfgVolatileWideFieldAccess

pub trait CommonCfgVolatileWideFieldAccess<'a, A> {
    // Required methods
    fn queue_desc(self) -> WideVolatilePtr<'a, le32, A::Restricted>
       where A: RestrictAccess<ReadWrite>;
    fn queue_driver(self) -> WideVolatilePtr<'a, le32, A::Restricted>
       where A: RestrictAccess<ReadWrite>;
    fn queue_device(self) -> WideVolatilePtr<'a, le32, A::Restricted>
       where A: RestrictAccess<ReadWrite>;
}

Common configuration structure

Required Methods

fn queue_desc(self) -> WideVolatilePtr<'a, le32, A::Restricted>
where A: RestrictAccess<ReadWrite>,

The driver writes the physical address of Device Area here. See section Basic Facilities of a Virtio Device / Virtqueues.

fn queue_driver(self) -> WideVolatilePtr<'a, le32, A::Restricted>
where A: RestrictAccess<ReadWrite>,

The driver writes the physical address of Device Area here. See section Basic Facilities of a Virtio Device / Virtqueues.

fn queue_device(self) -> WideVolatilePtr<'a, le32, A::Restricted>
where A: RestrictAccess<ReadWrite>,

The driver writes the physical address of Device Area here. See section Basic Facilities of a Virtio Device / Virtqueues.